pyproject.toml
[project]
name = "first-drafter"
version = "0.0.1"
description = "A Slack-based chatbot for legal research and drafting tasks"
authors = [
    { name = "Masayuki Tani", email = "masayuki.tani@pro-innovation.net" }
]
license = "MIT"
readme = "README.md"
requires-python = ">=3.12.2"
package-mode = false
dependencies = [
    "langchain==0.2.7",
    "langchain-core==0.2.18",
    "langchain-community==0.2.7",
    "langchain-anthropic==0.1.20",
    "langchain-chroma==0.1.2",
    "langchain-openai==0.1.16",
    "langgraph==0.1.8",
    "momento==1.21.0",
    "pdfplumber==0.11.2",
    "pypdf==4.3.0",
    "python-dotenv>=1.0.1",
    "slack-bolt==1.19.1",
]
